+++ cmd/livesim2/app/routes.go
@@ -56,13 +56,18 @@ func (s *Server) Routes(ctx context.Context) error {
 	s.Router.MethodFunc("HEAD", "/static/*", s.embeddedStaticHandlerFunc)
 	s.Router.MethodFunc("GET", "/reqcount", s.reqCountHandlerFunc)
 	s.Router.MethodFunc("OPTIONS", "/*", s.optionsHandlerFunc)
-	s.Router.Handle("/player/*", createReversePlayerProxy("/player", "https://reference.dashif.org/xxx"))
+	s.Router.Handle("/player/*", createReversePlayerProxy("/player", s.Cfg.PlayURL))
 	s.Router.MethodFunc("GET", "/", s.indexHandlerFunc)
+	s.Router.MethodFunc("POST", "/*", s.laURLHandlerFunc)
 	// LiveRouter is mounted at /livesim2
 	s.LiveRouter.MethodFunc("GET", "/*", s.livesimHandlerFunc)
 	s.LiveRouter.MethodFunc("HEAD", "/*", s.livesimHandlerFunc)
+	s.Router.MethodFunc("OPTIONS", "/*", s.optionsHandlerFunc)
+	s.LiveRouter.MethodFunc("POST", "/*", s.laURLHandlerFunc)
 	// VodRouter is mounted at /vod
-	s.VodRouter.HandleFunc("/*", s.vodHandlerFunc)
+	s.VodRouter.MethodFunc("GET", "/*", s.vodHandlerFunc)
+	s.VodRouter.MethodFunc("HEAD", "/*", s.vodHandlerFunc)
+	s.VodRouter.MethodFunc("OPTIONS", "/*", s.optionsHandlerFunc)
 	// Redirect /livesim to /livesim2 and /livesim-chunked for backwards compatibility
 	s.Router.MethodFunc("GET", "/livesim/*", redirect("/livesim", "/livesim2"))
 	s.Router.MethodFunc("GET", "/livesim-chunked/*", redirect("/livesim-chunked", "/livesim2"))